wip: add team builder
