+++ app/public/src/pages/component/bot-builder/bot-logic.ts
@@ -1,0 +1,92 @@
+import { PkmIndex, Pkm } from "../../../../../types/enum/Pokemon";
+
+export const DEFAULT_BOT_STATE = {
+    steps: [
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      },
+      {
+        roundsRequired: 2,
+        board: []
+      }
+    ],
+    avatar: PkmIndex[Pkm.DITTO],
+    author: "",
+    elo: 1200,
+    name: Pkm.DITTO,
+    id: ""
+  }
+
