--- app/public/src/pages/component/bot-builder/builder-synergies.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-import Synergies from "../synergy/synergies"
-import { useAppSelector } from "../../../hooks"
-
-export default function BuilderSynergies() {
-  const synergies = useAppSelector((state) => state.lobby.synergies)
-  return <Synergies synergies={synergies} />
-}
