+++ app/public/src/pages/component/bot-builder/pokemon-picker.tsx
@@ -1,99 +1,87 @@
-import React from "react"
+import React, { useState } from "react"
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs"
 import { Pkm, PkmIndex } from "../../../../../types/enum/Pokemon"
-import CSS from "csstype"
 import PRECOMPUTED_TYPE_POKEMONS_ALL from "../../../../../models/precomputed/type-pokemons-all.json"
 import { Item } from "../../../../../types/enum/Item"
 import { getPortraitSrc } from "../../../utils"
-import { DetailledPkm, Emotion } from "../../../../../types"
+import { PkmWithConfig, Emotion } from "../../../../../types"
 import SynergyIcon from "../icons/synergy-icon"
 import { Synergy } from "../../../../../types/enum/Synergy"
+import { GamePokemonDetail } from "../game/game-pokemon-detail"
+import ReactTooltip from "react-tooltip"
 import PokemonFactory from "../../../../../models/pokemon-factory"
-import { Rarity } from "../../../../../types/enum/Game"
-
-const pokemonPoolStyle: CSS.Properties = {
-  display: "flex",
-  flexWrap: "wrap",
-  margin: "10px",
-  marginTop: "0px"
-}
-
-const imgStyle: CSS.Properties = {
-  width: "40px",
-  height: "40px",
-  imageRendering: "pixelated",
-  cursor: "var(--cursor-hover)"
-}
-
-const cursorStyle: CSS.Properties = {
-  cursor: "var(--cursor-hover)"
-}
 
 export default function PokemonPicker(props: {
-  selectEntity: React.Dispatch<React.SetStateAction<DetailledPkm | Item>>
+  selectEntity: React.Dispatch<React.SetStateAction<PkmWithConfig | Item>>
 }) {
+  const tabs = [...Object.keys(PRECOMPUTED_TYPE_POKEMONS_ALL), "none"]
+  const pokemonsPerTab = tabs.map((t) =>
+    t === "none" ? [Pkm.KECLEON, Pkm.ARCEUS] : PRECOMPUTED_TYPE_POKEMONS_ALL[t]
+  )
+  const [hoveredPokemon, setHoveredPokemon] = useState<Pkm>()
+
+  function handleOnDragStart(e: React.DragEvent, name: Pkm) {
+    e.dataTransfer.setData("pokemon", name)
+  }
+
   return (
-    <Tabs className="nes-container" style={pokemonPoolStyle}>
+    <Tabs className="nes-container" id="pokemon-picker">
       <TabList>
-        {Object.keys(PRECOMPUTED_TYPE_POKEMONS_ALL).map((t) => {
+        {tabs.map((t) => {
           return (
-            <Tab style={cursorStyle} key={t}>
-              <SynergyIcon type={t as Synergy} />
+            <Tab key={t}>
+              {t === "none" ? "?" : <SynergyIcon type={t as Synergy} />}
             </Tab>
           )
         })}
-        <Tab style={cursorStyle} key="none">
-          <img
-            src="assets/unown/unown-qm.png"
-            alt="?"
-            className="unown-icon"
-            style={{ margin: "-28px" }}
-          />
-        </Tab>
       </TabList>
 
-      {Object.keys(PRECOMPUTED_TYPE_POKEMONS_ALL).map((key) => {
-        const pokemons = PRECOMPUTED_TYPE_POKEMONS_ALL[key]
+      {pokemonsPerTab.map((pokemons, i) => {
         return (
-          <TabPanel key={key} style={{ display: "flex", flexWrap: "wrap" }}>
+          <TabPanel
+            key={"pokemons-tab-" + i}
+            style={{ display: "flex", flexWrap: "wrap" }}
+          >
             {pokemons.map((pkm) => {
               return (
                 <div
+                  className="pokemon"
                   onClick={() => {
                     props.selectEntity({
-                      pkm: pkm,
+                      name: pkm,
                       emotion: Emotion.NORMAL,
                       shiny: false
                     })
                   }}
-                  key={`${pkm}`}
+                  onMouseOver={() => {
+                    setHoveredPokemon(pkm)
+                  }}
+                  key={pkm}
+                  data-tip
+                  data-for="pokemon-detail"
+                  draggable
+                  onDragStart={(e) => handleOnDragStart(e, pkm)}
                 >
-                  <img style={imgStyle} src={getPortraitSrc(PkmIndex[pkm])} />
+                  <img src={getPortraitSrc(PkmIndex[pkm])} />
                 </div>
               )
             })}
+            {hoveredPokemon && (
+              <ReactTooltip
+                id="pokemon-detail"
+                className="customeTheme game-pokemon-detail-tooltip"
+                effect="float"
+                place="top"
+                offset={{ top: 20 }}
+              >
+                <GamePokemonDetail
+                  pokemon={PokemonFactory.createPokemonFromName(hoveredPokemon)}
+                />
+              </ReactTooltip>
+            )}
           </TabPanel>
         )
       })}
-
-      <TabPanel key="none" style={{ display: "flex", flexWrap: "wrap" }}>
-        {[Pkm.KECLEON, Pkm.ARCEUS].map((pkm) => {
-          return (
-            <div
-              onClick={() => {
-                props.selectEntity({
-                  pkm: pkm,
-                  emotion: Emotion.NORMAL,
-                  shiny: false
-                })
-              }}
-              key={`${pkm}`}
-            >
-              <img style={imgStyle} src={getPortraitSrc(PkmIndex[pkm])} />
-            </div>
-          )
-        })}
-      </TabPanel>
     </Tabs>
   )
 }
