+++ app/public/src/pages/component/bot-builder/team-builder.tsx
@@ -1,430 +1,129 @@
-import React, { useState } from "react"
+import React, { useMemo, useState } from "react"
 import { Item } from "../../../../../types/enum/Item"
-import { Pkm, PkmFamily } from "../../../../../types/enum/Pokemon"
+import { Pkm } from "../../../../../types/enum/Pokemon"
 import PokemonFactory from "../../../../../models/pokemon-factory"
 import SelectedEntity from "./selected-entity"
-import ModalMenu from "./modal-menu"
 import ItemPicker from "./item-picker"
 import PokemonPicker from "./pokemon-picker"
 import TeamEditor from "./team-editor"
-import ReactTooltip from "react-tooltip"
-import { IBot, IStep } from "../../../../../models/mongo-models/bot-v2"
-import CSS from "csstype"
-import { produce } from "immer"
-import { useAppSelector, useAppDispatch } from "../../../hooks"
-import { createBot, requestBotList } from "../../../stores/NetworkStore"
-import { setBotCreatorSynergies } from "../../../stores/LobbyStore"
-import BuilderSynergies from "./builder-synergies"
+import { IDetailledPokemon } from "../../../../../models/mongo-models/bot-v2"
 import { Synergy } from "../../../../../types/enum/Synergy"
-import { BotGuideButton } from "../buttons/bot-guide-button"
-import {
-  DetailledPkm,
-  Emotion,
-  ModalMode,
-  ReadWriteMode
-} from "../../../../../types"
-import { PkmIndex } from "../../../../../types/enum/Pokemon"
-import { useTranslation } from "react-i18next"
-
-const buttonsStyle: CSS.Properties = {
-  left: "10px",
-  position: "absolute",
-  display: "flex"
-}
-
-const buttonStyle: CSS.Properties = {
-  marginLeft: "10px",
-  marginTop: "10px",
-  marginRight: "10px"
-}
-
-const bottomContainerStyle: CSS.Properties = {
-  display: "flex",
-  width: "87%",
-  position: "absolute",
-  bottom: "0px",
-  right: "0px"
-}
-
-const DEFAULT_BOT_STATE = {
-  steps: [
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    },
-    {
-      roundsRequired: 2,
-      board: []
-    }
-  ],
-  avatar: PkmIndex[Pkm.DITTO],
-  author: "",
-  elo: 1200,
-  name: Pkm.DITTO,
-  id: ""
-}
+import { PkmWithConfig, Emotion } from "../../../../../types"
+import Synergies from "../synergy/synergies"
+import "./team-builder.css"
+import { computeSynergies } from "../../../../../models/colyseus-models/synergies"
 
 export default function TeamBuilder() {
-  const { t } = useTranslation()
-  const dispatch = useAppDispatch()
-  const [step, setStep] = useState<number>(0)
-  const [copyStep, setCopyStep] = useState<IStep | undefined>(undefined)
-  const [bot, setBot] = useState<IBot>(DEFAULT_BOT_STATE)
-  const [entity, setEntity] = useState<Item | DetailledPkm>({
-    pkm: Pkm.DEFAULT,
+  const [selection, setSelection] = useState<Item | PkmWithConfig>({
+    name: Pkm.DEFAULT,
     shiny: false,
     emotion: Emotion.NORMAL
   })
-  const [mode, setMode] = useState<ReadWriteMode>(ReadWriteMode.WRITE)
-  const [modalMode, setModalMode] = useState<ModalMode>(ModalMode.IMPORT)
-  const [modalBoolean, setModalBoolean] = useState<boolean>(false)
 
-  const pastebinUrl: string = useAppSelector((state) => state.lobby.pastebinUrl)
-  const botData: IBot = useAppSelector((state) => state.lobby.botData)
-  const bots = useAppSelector((state) => state.lobby.botList)
+  const [board, setBoard] = useState<IDetailledPokemon[]>([])
+  const [selectedPokemon, setSelectedPokemon] = useState<IDetailledPokemon>()
 
-  if (bots.length === 0) {
-    dispatch(requestBotList())
+  const synergies: [Synergy, number][] = useMemo(() => {
+    const map = computeSynergies(
+      board.map((p) => {
+        const pkm = PokemonFactory.createPokemonFromName(p.name, {
+          selectedEmotion: p.emotion,
+          selectedShiny: p.shiny
+        })
+        pkm.positionX = p.x
+        pkm.positionY = p.y
+        p.items.forEach((item) => {
+          pkm.items.add(item)
+        })
+        return pkm
+      })
+    )
+    return [...map.entries()]
+  }, [board])
+
+  function addPokemon(x: number, y: number, pkm: PkmWithConfig) {
+    const i = board.findIndex((p) => p.x === x && p.y === y)
+    if (i >= 0) board.splice(i, 1)
+    const newPokemon: IDetailledPokemon = { ...pkm, x, y, items: [] }
+    setSelectedPokemon(newPokemon)
+    setBoard([...board, newPokemon])
   }
 
-  function updateSynergies(i: number) {
-    const newSynergies = new Map<Synergy, number>()
-    ;(Object.keys(Synergy) as Synergy[]).forEach((s) => {
-      newSynergies.set(s, 0)
-    })
-    const pokemonNames = new Array<Pkm>()
-    bot.steps[i].board.forEach((pkm) => {
-      const family = PkmFamily[pkm.name]
-      const pkmTypes = PokemonFactory.createPokemonFromName(pkm.name).types
-      if (!pokemonNames.includes(family)) {
-        pokemonNames.push(family)
-        pkmTypes.forEach((type: Synergy) => {
-          const v = newSynergies.get(type)
-          if (v) {
-            newSynergies.set(type, v + 1)
-          } else {
-            newSynergies.set(type, 1)
-          }
-        })
+  function addItem(x: number, y: number, item: Item) {
+    const p = board.find((p) => p.x === x && p.y === y)
+    if (p && p.items.length < 3) {
+      p.items.push(item)
+    } else if (p && p.items.length >= 3) {
+      p.items = [item]
+    }
+    setBoard([...board])
+  }
+
+  function handleEditorClick(x: number, y: number, rightClick: boolean) {
+    const pokemonOnCell = board.find((p) => p.x === x && p.y === y)
+    if (rightClick) {
+      setBoard(board.filter((p) => p !== pokemonOnCell))
+      if (
+        selectedPokemon &&
+        selectedPokemon.x === x &&
+        selectedPokemon.y === y
+      ) {
+        setSelectedPokemon(undefined)
       }
-    })
-    dispatch(setBotCreatorSynergies(newSynergies))
-  }
-
-  function write(x: number, y: number) {
-    if (Object.values(Pkm).includes((entity as DetailledPkm).pkm)) {
-      writePokemon(x, y)
-    }
-    if (Object.keys(Item).includes(entity as Item)) {
-      writeItem(x, y)
+    } else if (pokemonOnCell) {
+      setSelection(pokemonOnCell)
+      setSelectedPokemon(pokemonOnCell)
+    } else if (Object.values(Pkm).includes((selection as PkmWithConfig).name)) {
+      addPokemon(x, y, selection as PkmWithConfig)
+    } else if (Object.keys(Item).includes(selection as Item)) {
+      addItem(x, y, selection as Item)
     }
   }
 
-  function writeItem(x: number, y: number) {
-    const potential = bot.steps[step].board.findIndex(
-      (p) => p.x == x && p.y == y
-    )
-    const e = entity as Item
-    if (potential >= 0) {
-      if (bot.steps[step].board[potential].items.length < 3) {
-        setBot(
-          produce((draft) => {
-            draft.steps[step].board[potential].items.push(e)
-          })
-        )
-      } else {
-        setBot(
-          produce((draft) => {
-            draft.steps[step].board[potential].items = [e]
-          })
-        )
+  function handleDrop(x: number, y: number, e: React.DragEvent) {
+    if (e.dataTransfer.getData("cell") != "") {
+      const [originX, originY] = e.dataTransfer.getData("cell").split(",")
+      const pkm = board.find((p) => p.x === +originX && p.y === +originY)
+      if (pkm) {
+        pkm.x = x
+        pkm.y = y
+        setBoard([...board])
       }
+    } else if (e.dataTransfer.getData("pokemon") != "") {
+      const pkm: PkmWithConfig = {
+        name: e.dataTransfer.getData("pokemon") as Pkm,
+        emotion: Emotion.NORMAL,
+        shiny: false
+      }
+      addPokemon(x, y, pkm)
+      setSelection(pkm)
+    } else if (e.dataTransfer.getData("item") != "") {
+      const item = e.dataTransfer.getData("item") as Item
+      addItem(x, y, item)
+      setSelection(item)
     }
   }
 
-  function writePokemon(x: number, y: number) {
-    const potential = bot.steps[step].board.findIndex(
-      (p) => p.x == x && p.y == y
-    )
-    const e = entity as DetailledPkm
-    if (potential >= 0) {
-      setBot(
-        produce((draft) => {
-          draft.steps[step].board[potential].name = e.pkm
-          draft.steps[step].board[potential].shiny = e.shiny
-          draft.steps[step].board[potential].emotion = e.emotion
-        })
-      )
-    } else {
-      setBot(
-        produce((draft) => {
-          draft.steps[step].board.push({
-            name: e.pkm,
-            emotion: e.emotion,
-            shiny: e.shiny,
-            x: x,
-            y: y,
-            items: []
-          })
-        })
-      )
+  function updateSelectedPokemon(pkm: PkmWithConfig) {
+    setSelection(pkm)
+    if (selectedPokemon != null) {
+      selectedPokemon.emotion = pkm.emotion
+      selectedPokemon.shiny = pkm.shiny
+      setBoard([...board])
     }
-    updateSynergies(step)
-  }
-
-  function erase(x: number, y: number) {
-    const potential = bot.steps[step].board.findIndex(
-      (p) => p.x == x && p.y == y
-    )
-    if (potential >= 0) {
-      setBot(
-        produce((draft) => {
-          draft.steps[step].board.splice(potential, 1)
-        })
-      )
-    }
-  }
-
-  function importBot(text: string) {
-    try {
-      const b: IBot = JSON.parse(text)
-      setBot(b)
-      updateSynergies(step)
-      setModalBoolean(false)
-    } catch (e) {
-      alert(e)
-    }
-  }
-
-  function create() {
-    dispatch(createBot(bot))
   }
 
   return (
-    <div className="bot-panel">
-      <div className="header" style={buttonsStyle}>
-        <button
-          style={buttonStyle}
-          onClick={() => {
-            setModalMode(ModalMode.IMPORT)
-            setModalBoolean(true)
-          }}
-          className="bubbly orange"
-        >
-          {t("import")}/{t("load")}
-        </button>
-        <button
-          style={buttonStyle}
-          onClick={() => {
-            setModalMode(ModalMode.EXPORT)
-            setModalBoolean(true)
-          }}
-          className="bubbly orange"
-        >
-          {t("export")}
-        </button>
-        <BotGuideButton />
-        <button
-          style={buttonStyle}
-          onClick={() =>
-            setMode(
-              mode == ReadWriteMode.WRITE
-                ? ReadWriteMode.ERASE
-                : ReadWriteMode.WRITE
-            )
-          }
-          className="bubbly green"
-          data-tip
-          data-for={"mode"}
-        >
-          {mode} {t("mode")}
-          <ReactTooltip id={"mode"} className="customeTheme" place="bottom">
-            <p>{t("current_edition_click")}</p>
-            <p>{t("write_mode_hint")}</p>
-            <p> {t("erase_mode_hint")}</p>
-          </ReactTooltip>
-        </button>
-        <button
-          style={buttonStyle}
-          onClick={() => {
-            setCopyStep(JSON.parse(JSON.stringify(bot.steps[step])))
-          }}
-          className="bubbly green"
-          data-tip
-          data-for={"copy"}
-        >
-          <ReactTooltip id={"copy"} className="customeTheme" place="bottom">
-            <p>{t("copy_current_step")}</p>
-            <p> {t("paste_current_step")}</p>
-          </ReactTooltip>
-          {t("copy_step")}
-        </button>
-        <button
-          style={buttonStyle}
-          onClick={() => {
-            if (copyStep) {
-              setBot(
-                produce((draft) => {
-                  draft.steps[step] = copyStep
-                })
-              )
-            }
-          }}
-          className="bubbly green"
-          data-tip
-          data-for={"paste"}
-        >
-          <ReactTooltip id={"paste"} className="customeTheme" place="bottom">
-            <p>{t("paste_current_step_click")}</p>
-            <p>{t("paste_current_step_hint")}</p>
-          </ReactTooltip>
-          {t("paste_step")}
-        </button>
-      </div>
-      <BuilderSynergies />
+    <div id="team-builder" className="nes-container">
+      <Synergies synergies={synergies} />
       <TeamEditor
-        step={step}
-        steps={bot.steps}
-        avatar={bot.avatar}
-        author={bot.author}
-        name={bot.name}
-        elo={bot.elo}
-        handleTabClick={(i: number) => {
-          updateSynergies(i)
-          setStep(i)
-        }}
-        handleEditorClick={(x, y) => {
-          mode == ReadWriteMode.WRITE ? write(x, y) : erase(x, y)
-        }}
-        handleEloChange={(e) => {
-          const v = parseInt(e.target.value)
-          if (!isNaN(v)) {
-            setBot(
-              produce((draft) => {
-                draft.elo = v
-              })
-            )
-          }
-        }}
-        handleAuthorChange={(e) => {
-          setBot(
-            produce((draft) => {
-              draft.author = e.target.value
-            })
-          )
-        }}
-        handleAvatarChange={(e) =>
-          setBot(
-            produce((draft) => {
-              draft.name = e.target.value
-              draft.avatar = `${PkmIndex[e.target.value]}/${Emotion.NORMAL}`
-            })
-          )
-        }
-        handleRoundsRequiredChange={(e) =>
-          setBot(
-            produce((draft) => {
-              const v = parseInt(e.target.value)
-              draft.steps[step].roundsRequired = v
-            })
-          )
-        }
+        board={board}
+        handleEditorClick={handleEditorClick}
+        handleDrop={handleDrop}
       />
-      <SelectedEntity entity={entity} selectEntity={setEntity} />
-
-      <div style={bottomContainerStyle}>
-        <ItemPicker selectEntity={setEntity} />
-        <PokemonPicker selectEntity={setEntity} />
-      </div>
-
-      <ModalMenu
-        modalBoolean={modalBoolean}
-        showModal={(mode: ModalMode) => {
-          setModalMode(mode)
-          setModalBoolean(true)
-        }}
-        bot={bot}
-        hideModal={() => {
-          setModalBoolean(false)
-        }}
-        modalMode={modalMode}
-        importBot={importBot}
-        pasteBinUrl={pastebinUrl}
-        createBot={create}
-        botData={botData}
-      />
+      <SelectedEntity entity={selection} onChange={updateSelectedPokemon} />
+      <ItemPicker selectEntity={setSelection} />
+      <PokemonPicker selectEntity={setSelection} />
     </div>
   )
 }
